requirements.txt
ctypes
winreg
numpy
pillow